fix pengguna

+++ pages/pengguna-hapus.php
@@ -7,21 +7,32 @@ if($db->connect_errno==0)
 {
     if($_POST["kode_user"])
     {
-        $kode_user = $_POST["kode_user"];
-
-        $sql = "DELETE FROM user
-                WHERE kode_user='$kode_user'";
-        $res = $db->query($sql);
-        if($res)
+        if($kode_user == $_SESSION["kode_user"])
         {
-            if($db->affected_rows>0)
-            {
-                header("Location: pengguna.php?success=3");
-            }
-        }
-        else
-            // echo "Error ".(DEVELOPMENT?":".$db->error:"");
-            header("Location: pengguna.php?error=foreignkey");
+            $sql = "DELETE FROM user
+                WHERE kode_user='$kode_user'";
+                $res = $db->query($sql);
+                if($res)
+                {
+                    if($db->affected_rows>0)
+                    {
+                        header("Location: pengguna.php?success=3");
+                    }
+                    else 
+                    {
+                        header("Location: pengguna.php?warning=perubahan");
+                    }
+                }
+                else
+                {
+                    $_SESSION["err_fk"] = "user ini sedang digunakan di tabel lain.";    
+                    header("Location: pengguna.php?error=foreignkey");
+                }
+        } else 
+        {
+            $_SESSION["err_fk"] = "tidak dapat menghapus user yang sedang login.";    
+            header("Location: pengguna.php?error=foreignkey");  
+        }     
     }   
 }
 else
